provide permissions in query endpoint
